Use Thread.yield on the SdlLoopRunner to have a single SdlAsyncLoopRunner
